requirements.txt
pytorch-lightning>=2.1.0
hydra-core>=1.3.2
hydra-optuna-sweeper>=1.2.0
torchmetrics>=1.2.0
albumentations>=1.3.1
torch>=2.0.0
torchvision>=0.15.0
wandb
matplotlib
seaborn
pandas
scikit-learn
optuna
pyrootutils
rich
fvcore
grad-cam
ultralytics
scipy
